refactor(frontend): decompose useSpecializedAnalysis into focused composables (#2372)
Split useSpecializedAnalysis.ts (583 lines, 5 concerns) into:
- useApiEndpointAnalysis.ts (129 lines)
- useConfigDuplicates.ts (53 lines)
- useEnvironmentAnalysis.ts (120 lines)
- useOwnershipAnalysis.ts (95 lines)
- useCrossLanguageAnalysis.ts (270 lines)

Original file preserved as 42-line facade that spreads all sub-composable
returns, maintaining full backward compatibility.
